parsers/requirements.txt
requests==2.9.1
tqdm==3.8.0
Pillow==3.1.1
click==6.6
pyquery==1.2.13
ipdb
Unidecode==0.04.19
python-slugify==1.2.1